+++ apps/funding/package.json
@@ -27,7 +27,7 @@
   "devDependencies": {
     "@eslint/js": "^9.25.0",
     "@tailwindcss/postcss": "^4.1.10",
-    "@types/node": "^24.0.3",
+    "@types/node": "^24.0.4",
     "@types/react": "^19.1.2",
     "@types/react-datepicker": "^6.2.0",
     "@types/react-dom": "^19.1.2",
