+++ apps/funding/src/components/main-button/mainButton.styles.ts
