+++ apps/funding/src/components/point-button/pointButton.styles.ts
